+++ lib/src/navigator_notification.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/widgets.dart';
+import 'package:implicit_navigator/implicit_navigator.dart';
 
 /// A notification that is dispatched whenever a [ImplicitNavigator]'s stack
 /// changes.